folder controller deletion fix

+++ app/controller/folders.php
@@ -76,18 +76,23 @@ class Folders extends Base {
 	}
 
 	public function delete( $f3, $params ) {
+		$f3->push( 'SESSION.messages', array( 'Last Position: ' . $_SERVER['HTTP_REFERER'] , 0 ) );
 		if( $f3->exists( 'POST.id' ) && is_numeric( $f3->get( 'POST.id' ) ) ) {
-			$folder = new \Model\Folder( $params['id'] );
-			$folder->delete();
-			$f3->SESSION->messages[] = array( "Folder was successfully deleted", 0 );
-			if( isset( $params['id'] ) )
-				$f3->reroute( '/folder/' . $params['id'] );
-			else
-				$f3->reroute( '/dashboard' );
+			$folder = new \Model\Folder( $f3->get( 'POST.id' ) );
+			if( $folder->dry() ) {
+				$f3->push( 'SESSION.messages', array( $f3->get( 'L.deletefoldererr'), 1 ) );
+			} else {
+				$folder->delete();
+				$f3->push( 'SESSION.messages', array( "Folder was successfully deleted", 0 ) );
+			}
 		} else {
-			$f3->SESSION->messages[] = array( "Could not delete folder: No valid folder id given", 1 );
-			$f3->reroute( '/dashboard' );
+			$f3->push( 'SESSION.messages', array( "Could not delete folder: No valid folder id given", 1 ) );
 		}
+
+//		if( isset( $params['id'] ) )
+//			$f3->reroute( '/folder/' . $params['id'] );
+//		else
+			$f3->reroute( '/dashboard' );
 	}
 
 	public function show( \Base $f3, $params ) {
